refactor: move register logic into a separate command
